Move endpoint configuration into graphql module
